Reorganize swift files
