+++ APIKit/Method.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+public enum Method: String {
+    case GET = "GET"
+    case POST = "POST"
+    case PUT = "PUT"
+    case HEAD = "HEAD"
+    case DELETE = "DELETE"
+    case PATCH = "PATCH"
+    case TRACE = "TRACE"
+    case OPTIONS = "OPTIONS"
+    case CONNECT = "CONNECT"
+}
